Breed Counting Farmer John's N N cows, conveniently numbered 1…N 1…N , are all standing in a row (they seem to do so often that it now takes very little prompting from Farmer John to line them up). Each cow has a breed ID: 1 for Holsteins, 2 for Guernseys, and 3 for Jerseys. Farmer John would like your help counting the number of cows of each breed that lie within certain intervals of the ordering.
INPUT FORMAT (file bcount.in):
The first line of input contains
N
N
and
Q
Q
(
1≤N≤100,000
1≤N≤100,000
,
1≤Q≤100,000
1≤Q≤100,000
).
The next
N
N
lines contain an integer that is either 1, 2, or 3, giving the breed ID of a single cow in the ordering.
The next
Q
Q
lines describe a query in the form of two integers
a,b
a,b
(
a≤b
a≤b
).

OUTPUT FORMAT (file bcount.out):
For each of the
Q
Q
queries
(a,b)
(a,b)
, print a line containing three numbers: the number of cows numbered
a…b
a…b
that are Holsteins (breed 1), Guernseys (breed 2), and Jerseys (breed 3).

SAMPLE INPUT:
6 3
2
1
1
3
2
1
1 6
3 3
2 4
SAMPLE OUTPUT:
3 2 1
1 0 0
2 0 1